ch05 : add `const` where applicable

+++ ch05/ch05/ex19.cpp
@@ -5,7 +5,7 @@ int main() {
     const int* ptr_a = &a;
 
     printf("a의 값 : %d\n", a);
-    printf("a의 주솟값 : %d\n", &a);
-    printf("ptr_a의 값 : %d\n", ptr_a);
+    printf("a의 주솟값 : %ld\n", (long) &a);
+    printf("ptr_a의 값 : %ld\n", (long) ptr_a);
     printf("ptr_a에 저장된 주소에 저장된 값 : %d\n", *ptr_a);
 }
